+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Changed
+
+- A diff of the actual vs. expected logs is now printed when there's a mismatch with `naacl-utils verify`.
+
 ## [v0.3.0](https://github.com/naacl2022-reproducibility-track/naacl-utils/releases/tag/v0.3.0) - 2022-02-22
 
 ### Changed